Aldaketak inda failed

+++ src/main/java/edu/mondragon/simulator_os/Management.java
@@ -143,7 +143,7 @@ public class Management {
         long startTime = System.currentTimeMillis();
         Thread.sleep(randomTime);
         long endTime = System.currentTimeMillis();
-        return (double)(endTime - startTime);
+        return endTime - startTime;
 
     }
